'1.0版更新'

+++ realtimeWin.py
@@ -272,7 +272,7 @@ class Shuiwen_Canvas(FigureCanvas):
 
     def line(self):
         from run.run2020.env2019.env.app.CloudConn.cane_harvester import ins_real_data_dict
-        self.data_y.append(int(ins_real_data_dict.get('water_temperature', 0)))
+        self.data_y.append(float(ins_real_data_dict.get('water_temperature', 0)))
         self.axes.plot(self.data_y, color='red')
         self.axes.grid(True)
         self.draw()
@@ -288,7 +288,7 @@ class Youya_Canvas(Shuiwen_Canvas):
 
     def line(self):
         from run.run2020.env2019.env.app.CloudConn.cane_harvester import ins_real_data_dict
-        self.data_y.append(int(ins_real_data_dict.get('oil_pressure', 0)))
+        self.data_y.append(float(ins_real_data_dict.get('oil_pressure', 0)))
         self.axes.plot(self.data_y, color='orange')
         self.axes.grid(True)
         self.draw()
@@ -304,7 +304,7 @@ class Fadongji_Canvas(Shuiwen_Canvas):
 
     def line(self):
         from run.run2020.env2019.env.app.CloudConn.cane_harvester import ins_real_data_dict
-        self.data_y.append(int(ins_real_data_dict.get('engine_speed', 0)))
+        self.data_y.append(float(ins_real_data_dict.get('engine_speed', 0)))
         self.axes.plot(self.data_y, color='yellow')
         self.axes.grid(True)
         self.draw()
@@ -320,7 +320,7 @@ class Genqieqi_Canvas(Shuiwen_Canvas):
 
     def line(self):
         from run.run2020.env2019.env.app.CloudConn.cane_harvester import torque_real_data_dict
-        self.data_y.append(int(torque_real_data_dict.get('1_torque_ch', 0)))
+        self.data_y.append(torque_real_data_dict.get('1_torque_ch', 0))
         self.axes.plot(self.data_y, color='black')
         self.axes.grid(True)
         self.draw()
@@ -335,7 +335,7 @@ class Shusonggun_Canvas(Shuiwen_Canvas):
 
     def line(self):
         from run.run2020.env2019.env.app.CloudConn.cane_harvester import torque_real_data_dict
-        self.data_y.append(int(torque_real_data_dict.get('2_torque_ch', 0)))
+        self.data_y.append(torque_real_data_dict.get('2_torque_ch', 0))
         self.axes.plot(self.data_y, color='blue')
         self.axes.grid(True)
         self.draw()
@@ -351,7 +351,7 @@ class Qieduandao_Canvas(Shuiwen_Canvas):
 
     def line(self):
         from run.run2020.env2019.env.app.CloudConn.cane_harvester import torque_real_data_dict
-        self.data_y.append(int(torque_real_data_dict.get('3_torque_ch', 0)))
+        self.data_y.append(torque_real_data_dict.get('3_torque_ch', 0))
         self.axes.plot(self.data_y, color='#4B7902')
         self.axes.grid(True)
         self.draw()
@@ -367,7 +367,7 @@ class Paifengji_Canvas(Shuiwen_Canvas):
 
     def line(self):
         from run.run2020.env2019.env.app.CloudConn.cane_harvester import torque_real_data_dict
-        self.data_y.append(int(torque_real_data_dict.get('4_torque_ch', 0)))
+        self.data_y.append(torque_real_data_dict.get('4_torque_ch', 0))
         self.axes.plot(self.data_y, color='#F00CCB')
         self.axes.grid(True)
         self.draw()
@@ -383,7 +383,7 @@ class Erjishusong_Canvas(Shuiwen_Canvas):
 
     def line(self):
         from run.run2020.env2019.env.app.CloudConn.cane_harvester import torque_real_data_dict
-        self.data_y.append(int(torque_real_data_dict.get('5_torque_ch', 0)))
+        self.data_y.append(torque_real_data_dict.get('5_torque_ch', 0))
         self.axes.plot(self.data_y, color='#A52A2A')
         self.axes.grid(True)
         self.draw()
